[stripe] fix definitions files and tests
