+++ types/stripe-v3/tsconfig.json
@@ -8,19 +8,16 @@
         "noImplicitAny": true,
         "noImplicitThis": true,
         "strictNullChecks": true,
-        "baseUrl": "../../",
+        "baseUrl": "../",
         "typeRoots": [
-            "../../"
+            "../"
         ],
         "types": [],
-        "paths": {
-            "stripe": ["stripe/v2"]
-        },
         "noEmit": true,
         "forceConsistentCasingInFileNames": true
     },
     "files": [
         "index.d.ts",
-        "stripe-tests.ts"
+        "stripe-v3-tests.ts"
     ]
 }
